Implemented service and repository

+++ Droneshop.Core/ApplicationService/Services/CustomerService.cs
@@ -55,6 +55,10 @@ namespace Droneshop.Core.ApplicationService.Services
 
         public Customer DeleteCustomer(int id)
         {
+            if (id < 1)
+            {
+                throw new ArgumentException("The Id entered has to be at least 1");
+            }
             return _customerRepository.DeleteCustomer(id);
         }
